Added form and popup to map.
